Move properties to a separate file to match project standards
